moved run update and run query to separate trait
